+++ src/Fruit.cpp
@@ -1,16 +1,16 @@
 #include "Fruit.hpp"
 
-Fruit::Fruit(int x, int y) noexcept : m_position(x, y), m_shape(8)
+Fruit::Fruit(float x, float y) noexcept : m_position(x, y), m_shape(8)
 {
     m_shape.setFillColor(sf::Color::Green);
 }
 
-sf::Vector2i const& Fruit::getPosition() const& noexcept
+sf::Vector2f const& Fruit::getPosition() const& noexcept
 {
     return m_position;
 }
 
-sf::Vector2i Fruit::getPosition() && noexcept
+sf::Vector2f Fruit::getPosition() && noexcept
 {
     return std::move(m_position);
 }
